requirements.txt
torch
torchvision
timm
pytorch-lightning
albumentations
opencv-python-headless
pandas
scikit-learn
tensorboard
pillow
